Move auth service to sentry.auth
