+++ src/sentry/auth/access.py
@@ -24,6 +24,7 @@ from django.contrib.auth.models import AnonymousUser
 
 from sentry import features, roles
 from sentry.auth.services.access.service import access_service
+from sentry.auth.services.auth import RpcAuthState, RpcMemberSsoState
 from sentry.auth.staff import is_active_staff
 from sentry.auth.superuser import get_superuser_scopes, is_active_superuser
 from sentry.auth.system import SystemToken, is_system_auth
@@ -37,7 +38,6 @@ from sentry.models.team import Team, TeamStatus
 from sentry.models.user import User
 from sentry.roles import organization_roles
 from sentry.roles.manager import OrganizationRole, TeamRole
-from sentry.services.hybrid_cloud.auth import RpcAuthState, RpcMemberSsoState
 from sentry.services.hybrid_cloud.organization import RpcTeamMember, RpcUserOrganizationContext
 from sentry.services.hybrid_cloud.organization.serial import summarize_member
 from sentry.services.hybrid_cloud.user import RpcUser
