+++ src/sentry/auth/services/access/impl.py
@@ -1,18 +1,18 @@
 from sentry import roles
 from sentry.auth.services.access.service import AccessService
+from sentry.auth.services.auth import RpcAuthIdentity, RpcAuthProvider
+from sentry.auth.services.auth.serial import (
+    serialize_auth_identity,
+    serialize_auth_identity_replica,
+    serialize_auth_provider,
+    serialize_auth_provider_replica,
+)
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.authidentityreplica import AuthIdentityReplica
 from sentry.models.authprovider import AuthProvider
 from sentry.models.authproviderreplica import AuthProviderReplica
 from sentry.models.organizationmember import OrganizationMember
 from sentry.models.organizationmembermapping import OrganizationMemberMapping
-from sentry.services.hybrid_cloud.auth import RpcAuthIdentity, RpcAuthProvider
-from sentry.services.hybrid_cloud.auth.serial import (
-    serialize_auth_identity,
-    serialize_auth_identity_replica,
-    serialize_auth_provider,
-    serialize_auth_provider_replica,
-)
 from sentry.services.hybrid_cloud.organization import RpcOrganizationMemberSummary
 from sentry.services.hybrid_cloud.user.service import user_service
