+++ src/sentry/auth/services/access/service.py
@@ -4,13 +4,13 @@ from datetime import timedelta
 from django.utils import timezone
 
 from sentry.auth import find_providers_requiring_refresh
-from sentry.hybridcloud.rpc import silo_mode_delegation
-from sentry.services.hybrid_cloud.auth import (
+from sentry.auth.services.auth import (
     RpcAuthIdentity,
     RpcAuthProvider,
     RpcAuthState,
     RpcMemberSsoState,
 )
+from sentry.hybridcloud.rpc import silo_mode_delegation
 from sentry.services.hybrid_cloud.organization import RpcOrganizationMemberSummary
 from sentry.silo.base import SiloMode
