+++ src/sentry/models/authprovider.py
@@ -64,7 +64,7 @@ class AuthProvider(ReplicatedControlModel):
     default_global_access = models.BooleanField(default=True)
 
     def handle_async_replication(self, region_name: str, shard_identifier: int) -> None:
-        from sentry.services.hybrid_cloud.auth.serial import serialize_auth_provider
+        from sentry.auth.services.auth.serial import serialize_auth_provider
         from sentry.services.hybrid_cloud.replica.service import region_replica_service
 
         serialized = serialize_auth_provider(self)
