+++ src/sentry/receivers/outbox/region.py
@@ -11,12 +11,12 @@ from typing import Any
 
 from django.dispatch import receiver
 
+from sentry.auth.services.auth import auth_service
 from sentry.models.authproviderreplica import AuthProviderReplica
 from sentry.models.organization import Organization
 from sentry.models.outbox import OutboxCategory, process_region_outbox
 from sentry.models.project import Project
 from sentry.receivers.outbox import maybe_process_tombstone
-from sentry.services.hybrid_cloud.auth import auth_service
 from sentry.services.hybrid_cloud.log import AuditLogEvent, UserIpEvent, log_rpc_service
 from sentry.services.hybrid_cloud.organization_mapping import organization_mapping_service
 from sentry.services.hybrid_cloud.organization_mapping.model import CustomerId
