+++ src/sentry/services/hybrid_cloud/replica/impl.py
@@ -4,6 +4,7 @@ from typing import Any
 from django.db import router, transaction
 from django.db.models import Q
 
+from sentry.auth.services.auth import RpcApiKey, RpcApiToken, RpcAuthIdentity, RpcAuthProvider
 from sentry.db.models import BaseModel, FlexibleForeignKey
 from sentry.db.models.fields.hybrid_cloud_foreign_key import HybridCloudForeignKey
 from sentry.db.models.outboxes import ReplicatedControlModel, ReplicatedRegionModel
@@ -35,12 +36,6 @@ from sentry.models.team import Team
 from sentry.models.teamreplica import TeamReplica
 from sentry.models.user import User
 from sentry.notifications.services import RpcExternalActor
-from sentry.services.hybrid_cloud.auth import (
-    RpcApiKey,
-    RpcApiToken,
-    RpcAuthIdentity,
-    RpcAuthProvider,
-)
 from sentry.services.hybrid_cloud.organization import RpcOrganizationMemberTeam, RpcTeam
 from sentry.services.hybrid_cloud.orgauthtoken.model import RpcOrgAuthToken
 from sentry.services.hybrid_cloud.replica.service import ControlReplicaService, RegionReplicaService
