+++ src/sentry/services/hybrid_cloud/replica/service.py
@@ -1,17 +1,12 @@
 import abc
 
+from sentry.auth.services.auth import RpcApiKey, RpcApiToken, RpcAuthIdentity, RpcAuthProvider
 from sentry.hybridcloud.rpc.resolvers import ByRegionName
 from sentry.hybridcloud.rpc.service import RpcService, regional_rpc_method, rpc_method
 from sentry.hybridcloud.services.control_organization_provisioning import (
     RpcOrganizationSlugReservation,
 )
 from sentry.notifications.services import RpcExternalActor
-from sentry.services.hybrid_cloud.auth import (
-    RpcApiKey,
-    RpcApiToken,
-    RpcAuthIdentity,
-    RpcAuthProvider,
-)
 from sentry.services.hybrid_cloud.organization import RpcOrganizationMemberTeam, RpcTeam
 from sentry.services.hybrid_cloud.orgauthtoken.model import RpcOrgAuthToken
 from sentry.silo.base import SiloMode
