+++ src/sentry/web/frontend/organization_auth_settings.py
@@ -13,10 +13,10 @@ from rest_framework.request import Request
 from sentry import audit_log, features, roles
 from sentry.auth import manager
 from sentry.auth.helper import AuthHelper
+from sentry.auth.services.auth import RpcAuthProvider, auth_service
 from sentry.models.authprovider import AuthProvider
 from sentry.models.organization import Organization
 from sentry.plugins.base import Response
-from sentry.services.hybrid_cloud.auth import RpcAuthProvider, auth_service
 from sentry.services.hybrid_cloud.organization import RpcOrganization, organization_service
 from sentry.tasks.auth import email_missing_links_control
 from sentry.utils.http import absolute_uri
