+++ tests/sentry/auth/test_superuser.py
@@ -8,6 +8,7 @@ from django.core import signing
 from django.test import override_settings
 from django.utils import timezone
 
+from sentry.auth.services.auth.model import RpcAuthState, RpcMemberSsoState
 from sentry.auth.superuser import (
     COOKIE_DOMAIN,
     COOKIE_HTTPONLY,
@@ -31,7 +32,6 @@ from sentry.auth.superuser import (
 from sentry.auth.system import SystemToken
 from sentry.middleware.placeholder import placeholder_get_response
 from sentry.middleware.superuser import SuperuserMiddleware
-from sentry.services.hybrid_cloud.auth.model import RpcAuthState, RpcMemberSsoState
 from sentry.testutils.cases import TestCase
 from sentry.testutils.helpers.datetime import freeze_time
 from sentry.testutils.helpers.options import override_options
