+++ tests/sentry/hybridcloud/test_replica.py
@@ -1,3 +1,4 @@
+from sentry.auth.services.auth.serial import serialize_auth_provider
 from sentry.hybridcloud.models import ApiKeyReplica, ExternalActorReplica
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.authidentityreplica import AuthIdentityReplica
@@ -6,7 +7,6 @@ from sentry.models.authproviderreplica import AuthProviderReplica
 from sentry.models.organizationmemberteamreplica import OrganizationMemberTeamReplica
 from sentry.models.outbox import outbox_context
 from sentry.models.teamreplica import TeamReplica
-from sentry.services.hybrid_cloud.auth.serial import serialize_auth_provider
 from sentry.services.hybrid_cloud.replica import region_replica_service
 from sentry.silo.base import SiloMode
 from sentry.testutils.factories import Factories
